+++ Cast/Events/VolumeChanged.swift
@@ -10,8 +10,14 @@ import Foundation
 
 /// Receiver message detailing a change in volume
 public struct VolumeChanged: Decodable {
+    /// The descrete volume
+    @available(*, introduced: 0.73.0, deprecated: 2.0.82)
+    public var volume: Int {
+        return Int(volumeLevel * 100)
+    }
+    
     /// The volume level
-    public let volume: Int
+    public let volumeLevel: Float
     
     /// If the volume is muted or not
     public let muted: Bool
@@ -21,13 +27,15 @@ public struct VolumeChanged: Decodable {
         let container = try decoder.container(keyedBy: BaseKeys.self)
         let nested = try container.nestedContainer(keyedBy: DataKeys.self, forKey: .data)
         
-        volume = try nested.decode(Int.self, forKey: .volume)
+        let percent = try nested.decode(Float.self, forKey: .volume)
+        volumeLevel = percent
         muted = try nested.decode(Bool.self, forKey: .muted)
     }
     
     enum BaseKeys: CodingKey {
         case data
     }
+    
     enum DataKeys: CodingKey {
         case volume
         case muted