+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 * `2.0.79` Release - [2.0.79](#2079)
 * `0.73.x` Releases - [0.73.0](#0730)
 
+## 2.0.82
+
+#### Bug Fixes
+* Fixed a parsing issue for the `VolumeChanged` event
+* Improved resilience on error message parsing 
+
 ## 2.0.79
 
 * Adopted new *Receiver 2.0* api.